A whole lot of changes to get compiler warning flags to go away. You're probably better off using size_t and ssize_t when dealing with C library calls anyways.

+++ src/bson.h
@@ -86,10 +86,10 @@ typedef struct {
 typedef struct {
     char *data;
     char *cur;
-    int dataSize;
+    size_t dataSize;
     bson_bool_t finished;
-    int stack[32];
-    int stackPos;
+    size_t stack[32];
+    size_t stackPos;
     int err; /**< Bitfield representing errors or warnings on this buffer */
     char *errstr; /**< A string representation of the most recent error or warning. */
 } bson;
@@ -557,7 +557,7 @@ void bson_init_size( bson *b, int size );
  * @return BSON_OK or BSON_ERROR with the bson error object set.
  *   Exits if allocation fails.
  */
-int bson_ensure_space( bson *b, const int bytesNeeded );
+int bson_ensure_space( bson *b, const size_t bytesNeeded );
 
 /**
  * Finalize a bson object.
@@ -744,7 +744,7 @@ int bson_append_code_w_scope( bson *b, const char *name, const char *code, const
  *
  * @return BSON_OK or BSON_ERROR.
  */
-int bson_append_code_w_scope_n( bson *b, const char *name, const char *code, int size, const bson *scope );
+int bson_append_code_w_scope_n( bson *b, const char *name, const char *code, size_t size, const bson *scope );
 
 /**
  * Append binary data to a bson.
@@ -757,7 +757,7 @@ int bson_append_code_w_scope_n( bson *b, const char *name, const char *code, int
  *
  * @return BSON_OK or BSON_ERROR.
  */
-int bson_append_binary( bson *b, const char *name, char type, const char *str, int len );
+int bson_append_binary( bson *b, const char *name, char type, const char *str, size_t len );
 
 /**
  * Append a bson_bool_t to a bson.
@@ -930,7 +930,7 @@ extern bson_printf_func bson_errprintf;
  *
  * @sa malloc(3)
  */
-void *bson_malloc( int size );
+void *bson_malloc( size_t size );
 
 /**
  * Changes the size of allocated memory and checks return value,
@@ -943,7 +943,7 @@ void *bson_malloc( int size );
  *
  * @sa realloc()
  */
-void *bson_realloc( void *ptr, int size );
+void *bson_realloc( void *ptr, size_t size );
 
 /**
  * Set a function for error handling.
